ft(two-factor-auth for sellers) implement a two factor authentication page for vendors implement OTP input field validation implement OTP validation and redirect the seller to dashboard on success
