+++ src/pages/Home.jsx
@@ -1,9 +1,9 @@
 import Header from '../components/Header';
-import ProductList from '../components/product/ProductList';
+// import ProductList from '../components/product/ProductList';
 
 const Home = () => {
 	return (
-		<div data-testid="homePage">
+		<div>
 			<Header />
 			{/* <ProductList /> */}
 		</div>
