+++ src/pages/LoginPage.jsx
@@ -36,7 +36,7 @@ const Login = () => {
 
 			if (user?.data?.others?.RoleId === 3) return navigate('/');
 
-			if (user?.loginOTP) return navigate('/vendor');
+			if (user?.hashedOTP) return navigate('/auth');
 		}, 9000);
 	}, [user]);
